Update util.h
glPrintf uses double

+++ student_template_A2024/util.h
@@ -3,7 +3,7 @@
 
 long map(long x, long in_min, long in_max, long out_min, long out_max);
 float mapf(float x, float in_min, float in_max, float out_min, float out_max);
-void glPrintf(int x, int y, int z, const char* format, ...);
+void glPrintf(double x, double y, double z, const char* format, ...);
 void open_gl_init(void);
 #endif
